requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
black==21.9b0
certifi==2021.5.30
click==8.0.1
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.7
django-bootstrap4==3.0.1
django-heroku==0.3.1
django-registration==3.2
django-registration-redux==2.9
djangorestframework==3.12.4
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
regex==2021.8.28
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
